Split pictogram password widget in two Avoids unneccesary reloading
