+++ lib/widgets/pictogram_password_widget.dart
@@ -1,31 +1,12 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:weekplanner/di.dart';
 import 'package:api_client/models/pictogram_model.dart';
+import 'package:weekplanner/di.dart';
+import 'package:weekplanner/widgets/paswordInputField.dart';
 import 'package:weekplanner/widgets/pictogram_image.dart';
 import 'package:api_client/api/api.dart';
-
 import 'giraf_notify_dialog.dart';
 
-///Widget to show possible pictograms for pictogram password
-///and the currently input pictograms
-class PictogramChoices extends StatefulWidget {
-  ///Widget with the possible pictograms in the code and the currently picked
-  /// pictograms in the code.
-
-  const PictogramChoices({Key key,
-    @required this.onPasswordChanged, @required this.api})
-      : super(key: key);
-
-  /// This function returns the new password every time the password has been
-  /// changed by adding or removing pictogram.
-  final Function onPasswordChanged;
-
-  final Api api;
-
-  @override
-  _PictogramChoiceState createState() => _PictogramChoiceState(api);
-}
 
 /// The pictograms to choose between for the code.
 /// If these are changed all previously made passwords will become unusable
@@ -34,85 +15,50 @@ const List<int> CHOSENPICTOGRAMS = <int>[1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
 /// The maximum width that is shared by both gridviews
 const double MAXWIDTH = 500;
 
-class _PictogramChoiceState extends State<PictogramChoices> {
-  _PictogramChoiceState(this._api);
-  /// Api to use for pictogram calls
-  final Api _api;
-  /// List of currently chosen pictograms for the code
-  List<PictogramModel> _inputCode;
+///Widget to show possible pictograms for pictogram password
+///and the currently input pictograms
+class PictogramChoices extends StatelessWidget {
+  ///Widget with the possible pictograms in the code and the currently picked
+  /// pictograms in the code.
 
-  /// Stream of usable pictograms
-  Stream<List<PictogramModel>> _pictogramChoices;
+  const PictogramChoices({Key key,
+    @required this.onPasswordChanged,
+    @required this.api})
+      : super(key: key);
 
-  @override
-  void initState() {
-    _inputCode = List<PictogramModel>.filled(4, null);
+  /// This function returns the new password every time the password has been
+  /// changed by adding or removing pictogram.
+  final Function onPasswordChanged;
 
-    _pictogramChoices = getStream();
-    super.initState();
-  }
+  /// Pictogram api passed as a parameter to allow for testing
+  final Api api;
 
-  ///Called when one of the possible pictograms are pressed
-  ///and adds it to the code
-  void addToPass(PictogramModel pictogram) {
-    final int index = _inputCode.indexOf(null);
-    if (index == -1) {
-      return;
-    }
-      _inputCode[index] = pictogram;
-      widget.onPasswordChanged(validateAndConvertPass());
-      //Reloads the widget with the new input
-      setState(() {});
-
-  }
-
-  ///Called when a pictogram is pressed in the code and removes the pressed icon
-  void removeFromPass(int index) {
-    _inputCode[index] = null;
-    widget.onPasswordChanged(validateAndConvertPass());
-    //Reloads the widget with the new code
-    setState(() {});
-  }
-
-  /// Validates the chosen pictograms and converts them into a string.
-  /// If the input is invalid, null is return. This means that one must
-  /// null-check when using this widget.
-  String validateAndConvertPass() {
-    String output = '';
-    bool valid = true;
-    passwordList().forEach((Widget w) {
-      if (w is PictogramImage) {
-        output += w.pictogram.id.toString();
-      } else {
-        valid = false;
-      }
-    });
-    return (valid == true) ? output : null;
-  }
-
+  /// Returns the list of possible pictograms for use in the password
+  /// as a stream
   Stream<List<PictogramModel>> getStream() async* {
     final List<PictogramModel> list =
-        List<PictogramModel>.filled(CHOSENPICTOGRAMS.length, null);
+    List<PictogramModel>.filled(CHOSENPICTOGRAMS.length, null);
     yield list;
-    try{
+    try {
       for (int i = 0; i < list.length; i++) {
         final int index = CHOSENPICTOGRAMS[i];
-        await for (final PictogramModel model in _api.pictogram.get(index)) {
+        await for (final PictogramModel model in api.pictogram.get(index)) {
           list[i] = model;
           yield list;
         }
       }
-    }catch(e){
-      showErrorMessage(e, context);
+    } catch (e) {
+      showErrorMessage(e, di.getDependency<BuildContext>());
+
     }
   }
-
+  /// Shows error message in case of any pictogram being unobtainable
   void showErrorMessage(Object error, BuildContext context) {
     showDialog<Center>(
 
       /// exception handler to handle web_api exceptions
         barrierDismissible: false,
-        context: context,
+        context:  context,
         builder: (BuildContext context) {
           return const GirafNotifyDialog(
               title: 'Fejl',
@@ -121,37 +67,13 @@ class _PictogramChoiceState extends State<PictogramChoices> {
         });
   }
 
-  List<Widget> passwordList() {
-    final List<Widget> password = List<Widget>.filled(4, null);
-    for (int i = 0; i < 4; i++) {
-      final PictogramModel pictogram = _inputCode[i];
-      Widget widget;
-      if (pictogram == null) {
-        widget = Container(
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(20),
-            color: const Color(0xFFe0dede),
-          ),
-        );
-      } else if (pictogram.id == -1){
-        // For testing purposes
-        widget = const Text(
-            'LoginPictogram',
-            //key: Key('LoginPictogram')
-        );
-      }
-      else
-      {
-        widget = PictogramImage(
-            pictogram: pictogram, onPressed: () => removeFromPass(i));
-      }
-      password[i] = widget;
-    }
-    return password;
-  }
-
   @override
   Widget build(BuildContext context) {
+    final Stream<List<PictogramModel>> _pictogramChoices = getStream();
+    final GlobalKey<PasswordInputFieldState> inputFieldKey = GlobalKey();
+    final PasswordInputField password = PasswordInputField(
+        key: inputFieldKey,
+        onPasswordChanged: onPasswordChanged);
     return Column(children: <Widget>[
       // Grid view with available pictograms
       Row(
@@ -182,18 +104,21 @@ class _PictogramChoiceState extends State<PictogramChoices> {
                               );
                             }
                             // This is here to allow for testing.
-                            else if (pictogram.id == -1){
+                            else if (pictogram.id == -1) {
                               return GestureDetector(
-                                onTap: () => addToPass(pictogram),
+                                onTap: () =>
+                                    inputFieldKey.
+                                    currentState.addToPass(pictogram),
                                 child: const Text('Error'),
                                 key: const Key('TestPictogram'),
                               );
                             }
-                            else
-                            {
+                            else {
                               return PictogramImage(
                                   pictogram: pictogram,
-                                  onPressed: () => addToPass(pictogram));
+                                  onPressed: () =>
+                                      inputFieldKey.
+                                      currentState.addToPass(pictogram));
                             }
                           }).toList());
                     } else {
@@ -207,26 +132,7 @@ class _PictogramChoiceState extends State<PictogramChoices> {
         height: 10,
       ),
       // Password grid view
-      Container(
-        color: Colors.grey,
-        width: MediaQuery.of(context).size.width,
-        height: 120,
-        key: const Key('InputPasswordContainer'),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Container(
-                constraints: const BoxConstraints(
-                    maxHeight: double.infinity, maxWidth: MAXWIDTH),
-                child: GridView.count(
-                    physics: const NeverScrollableScrollPhysics(),
-                    shrinkWrap: true,
-                    crossAxisCount: 4,
-                    crossAxisSpacing: 20,
-                    children: passwordList())),
-          ],
-        ),
-      )
+      password
     ]);
   }
 }
